pyproject.toml
[project]
name = "tgbot"
version = "0.1.0"
description = "Telegram in-channel Q&A bot with replayable ingestion pipeline"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  # Bot runtime
  "aiogram==3.24.0",
  # Userbot/backfill
  "Telethon==1.42.0",
  # Config
  "pydantic-settings==2.12.0",
  # Storage
  "SQLAlchemy==2.0.45",
  "alembic==1.18.0",
  "psycopg[binary]==3.2.3",
  # Ops API
  "fastapi==0.128.0",
  "uvicorn==0.35.0",
  # Utils
  "python-dotenv==1.0.1",
]

[build-system]
requires = ["setuptools>=70", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[project.optional-dependencies]
dev = [
  "pytest==8.3.4",
  "ruff==0.8.2",
  "mypy==1.13.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]

